[PDA-760] move modal to separate component
