+++ packages/app/src/renderer/modules/overview/overview-ui.component.tsx
@@ -15,6 +15,7 @@ import Backup from "Renderer/components/rest/overview/backup/backup.component"
 import { noop } from "Renderer/utils/noop"
 import { PhoneUpdate } from "Renderer/models/phone-update/phone-update.interface"
 import { AppSettings } from "App/main/store/settings.interface"
+import CollectingDataWrapper from "App/contacts/components/collecting-data-modal/collecting-data-wrapper.component"
 
 const PhoneInfo = styled(Phone)`
   grid-area: Phone;
@@ -56,6 +57,7 @@ interface OverviewUIProps {
   readonly onOpenBackupModal: () => void
   readonly onOpenBackupRestorationModal: () => void
   toggleDevMode?: () => void
+  setCollectingData: (option: AppSettings["appCollectingData"]) => void
 }
 
 const OverviewUI: FunctionComponent<
@@ -86,6 +88,8 @@ const OverviewUI: FunctionComponent<
   pureOsDownloaded,
   simCards,
   toggleDevMode,
+  appCollectingData,
+  setCollectingData,
 }) => (
   <OverviewWrapper>
     <PhoneInfo
@@ -115,6 +119,10 @@ const OverviewUI: FunctionComponent<
       onBackupCreate={onOpenBackupModal}
       onBackupRestore={onOpenBackupRestorationModal}
       language={language}
+    />
+    <CollectingDataWrapper
+      appCollectingData={appCollectingData}
+      setCollectingData={setCollectingData}
     />
   </OverviewWrapper>
 )
